chore: split nav for ssr, reorganized icons
