+++ components/icons.tsx
@@ -1,6 +1,27 @@
-import React from 'react';
+export const MenuIcon = (
+  props: React.JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>,
+) => {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <line x1="4" x2="20" y1="12" y2="12" />
+      <line x1="4" x2="20" y1="6" y2="6" />
+      <line x1="4" x2="20" y1="18" y2="18" />
+    </svg>
+  );
+};
 
-const EnvelopeLogo = ({ className = '' }) => {
+export const EnvelopeLogo = ({ className = '' }) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -11,5 +32,3 @@ const EnvelopeLogo = ({ className = '' }) => {
     </svg>
   );
 };
-
-export default EnvelopeLogo;
