+++ components/layout/desktop-nav.tsx
@@ -1,0 +1,44 @@
+'use client';
+
+import { usePathname } from 'next/navigation';
+import Link from 'next/link';
+import {
+  NavigationMenu,
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuList,
+  navigationMenuTriggerStyle,
+} from '@/components/ui/navigation-menu';
+import { cn } from '@/lib/utils';
+
+export function DesktopNav({
+  links,
+}: {
+  links: { href: string; label: string }[];
+}) {
+  const pathname = usePathname();
+
+  return (
+    <NavigationMenu>
+      <NavigationMenuList className="hidden sm:flex">
+        {links.map((link) => (
+          <NavigationMenuItem key={link.href}>
+            <Link href={link.href} legacyBehavior passHref>
+              <NavigationMenuLink
+                className={cn(
+                  navigationMenuTriggerStyle(),
+                  'font-semibold',
+                  pathname === link.href
+                    ? 'underline underline-offset-4 text-primary'
+                    : 'dark:text-[var(--custom-dark-font-color)]',
+                )}
+              >
+                {link.label}
+              </NavigationMenuLink>
+            </Link>
+          </NavigationMenuItem>
+        ))}
+      </NavigationMenuList>
+    </NavigationMenu>
+  );
+}
